% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{all}
\alias{all}
\title{Returns a dataset containing raw size-frequency data for all categories of species}
\usage{
all()
}
\description{
Returns a dataset containing raw size-frequency data for all categories of species
}
