% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{SF.STD.all}
\alias{SF.STD.all}
\title{Returns a dataset containing standardized size-frequency data for all categories of species}
\usage{
SF.STD.all()
}
\description{
Returns a dataset containing standardized size-frequency data for all categories of species
}
