% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SF.STD.TROP}
\alias{SF.STD.TROP}
\title{The 'standardized' monthly size-frequency dataset for tropical tunas}
\format{
\describe{
  \item{YEAR}{}
  \item{QUARTER}{}
  \item{MONTH_START}{}
  \item{MONTH_END}{}
  \item{FISHING_GROUND_CODE}{}
  \item{FISHING_GROUND}{}
  \item{FLEET_CODE}{}
  \item{FLEET}{}
  \item{FISHERY_TYPE_CODE}{}
  \item{FISHERY_TYPE}{}
  \item{FISHERY_GROUP_CODE}{}
  \item{FISHERY_GROUP}{}
  \item{FISHERY_CODE}{}
  \item{FISHERY}{}
  \item{GEAR_CODE}{}
  \item{GEAR}{}
  \item{SCHOOL_TYPE_CODE}{}
  \item{IUCN_STATUS_CODE}{}
  \item{IUCN_STATUS}{}
  \item{SPECIES_WP_CODE}{}
  \item{SPECIES_WP}{}
  \item{SPECIES_GROUP_CODE}{}
  \item{SPECIES_GROUP}{}
  \item{SPECIES_CATEGORY_CODE}{}
  \item{SPECIES_CATEGORY}{}
  \item{SPECIES_CODE}{}
  \item{SPECIES}{}
  \item{SPECIES_SCIENTIFIC}{}
  \item{SPECIES_FAMILY}{}
  \item{SPECIES_ORDER}{}
  \item{IS_IOTC_SPECIES}{}
  \item{IS_SPECIES_AGGREGATE}{}
  \item{IS_SSI}{}
  \item{MEASURE_TYPE_CODE}{}
  \item{MEASURE_TYPE}{}
  \item{WEIGHT}{}
  \item{SEX_CODE}{}
  \item{CLASS_LOW}{}
  \item{CLASS_HIGH}{}
  \item{FISH_COUNT}{}
  \item{RAISING}{}
  \item{RAISE_CODE}{}
}
}
\usage{
SF.STD.TROP
}
\description{
The 'standardized' monthly size-frequency dataset for tropical tunas
}
\keyword{datasets}
