% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{STD.all}
\alias{STD.all}
\title{Returns a dataset containing standardized size-frequency data for all categories of species}
\usage{
STD.all()
}
\description{
Returns a dataset containing standardized size-frequency data for all categories of species
}
